+++ demo-extended/src/main/java/com/novoda/downloadmanager/demo/extended/batches/BeardBatch.java
@@ -1,14 +1,14 @@
 package com.novoda.downloadmanager.demo.extended.batches;
 
-import com.novoda.downloadmanager.demo.extended.Download;
+import com.novoda.downloadmanager.demo.extended.BeardDownload;
 import com.novoda.downloadmanager.lib.DownloadManager;
 
 /**
  * Model object to encapsulate data from the Batches table.
- * The difference between this and {@link Download} is that the batch doesn't
+ * The difference between this and {@link BeardDownload} is that the batch doesn't
  * contain the file path as it represents a batch of downloads.
  */
-public class Batch {
+public class BeardBatch {
     private final int id;
     private final String title;
     private final int status;
@@ -16,7 +16,7 @@ public class Batch {
     private final long currentBytes;
     private final String extraData;
 
-    public Batch(int id, String title, int status, long totalBytes, long currentBytes, String extraData) {
+    public BeardBatch(int id, String title, int status, long totalBytes, long currentBytes, String extraData) {
         this.id = id;
         this.title = title;
         this.status = status;