+++ demo-extended/src/main/java/com/novoda/downloadmanager/demo/extended/batches/QueryForBatchesAsyncTask.java
@@ -10,7 +10,7 @@ import java.lang.ref.WeakReference;
 import java.util.ArrayList;
 import java.util.List;
 
-public class QueryForBatchesAsyncTask extends AsyncTask<BatchQuery, Void, List<Batch>> {
+public class QueryForBatchesAsyncTask extends AsyncTask<BatchQuery, Void, List<BeardBatch>> {
 
     private final DownloadManager downloadManager;
     private final WeakReference<Callback> weakCallback;
@@ -25,9 +25,9 @@ public class QueryForBatchesAsyncTask extends AsyncTask<BatchQuery, Void, List<B
     }
 
     @Override
-    protected List<Batch> doInBackground(BatchQuery... params) {
+    protected List<BeardBatch> doInBackground(BatchQuery... params) {
         Cursor cursor = downloadManager.query(params[0]);
-        List<Batch> batches = new ArrayList<>();
+        List<BeardBatch> beardBatches = new ArrayList<>();
         try {
             while (cursor.moveToNext()) {
                 int id = cursor.getInt(cursor.getColumnIndexOrThrow(DownloadManager.COLUMN_ID));
@@ -36,25 +36,25 @@ public class QueryForBatchesAsyncTask extends AsyncTask<BatchQuery, Void, List<B
                 long totalBytes = cursor.getLong(cursor.getColumnIndexOrThrow(DownloadManager.COLUMN_BATCH_TOTAL_SIZE_BYTES));
                 long currentBytes = cursor.getLong(cursor.getColumnIndexOrThrow(DownloadManager.COLUMN_BATCH_CURRENT_SIZE_BYTES));
                 String extraData = cursor.getString(cursor.getColumnIndexOrThrow(DownloadManager.COLUMN_EXTRA_DATA));
-                batches.add(new Batch(id, title, status, totalBytes, currentBytes, extraData));
+                beardBatches.add(new BeardBatch(id, title, status, totalBytes, currentBytes, extraData));
             }
         } finally {
             cursor.close();
         }
-        return batches;
+        return beardBatches;
     }
 
     @Override
-    protected void onPostExecute(List<Batch> batches) {
-        super.onPostExecute(batches);
+    protected void onPostExecute(List<BeardBatch> beardBatches) {
+        super.onPostExecute(beardBatches);
         Callback callback = weakCallback.get();
         if (callback == null) {
             return;
         }
-        callback.onQueryResult(batches);
+        callback.onQueryResult(beardBatches);
     }
 
     public interface Callback {
-        void onQueryResult(List<Batch> batches);
+        void onQueryResult(List<BeardBatch> beardBatches);
     }
 }